Move getProviderFromUrl and trackOAuthCompletion to analytics-providers
