+++ plugins/login-resources/package.json
@@ -49,6 +49,7 @@
     "@hcengineering/setting": "^0.6.17",
     "@hcengineering/theme": "^0.6.5",
     "@hcengineering/analytics": "^0.6.0",
-    "@hcengineering/account-client": "^0.6.0"
+    "@hcengineering/account-client": "^0.6.0",
+    "@hcengineering/analytics-providers": "^0.6.0"
   }
 }
