+++ plugins/login-resources/src/components/Join.svelte
@@ -19,7 +19,8 @@
   import { checkJoined, join, setLoginInfo, signUpJoin } from '../utils'
   import Form from './Form.svelte'
 
-  import { Analytics, signupStore } from '@hcengineering/analytics'
+  import { Analytics } from '@hcengineering/analytics'
+  import { signupStore } from '@hcengineering/analytics-providers'
   import { logIn, workbenchId } from '@hcengineering/workbench'
   import { onMount } from 'svelte'
   import { loginAction, recoveryAction } from '../actions'
